fastapi/rinha-backend-2023/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.103.1
h11==0.14.0
idna==3.4
mypy==1.5.1
mypy-extensions==1.0.0
psycopg2-binary==2.9.7
pydantic==1.10.12
pydantic_core==2.6.3
python-decouple==3.8
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.7.1
ulid==1.1
uvicorn==0.23.2
